requirements.txt
# Requirements for OpsAI
# Python dependencies for AI-powered operations platform
#
# Author: Ayush
# GitHub: https://github.com/pheonix-19
# Copyright (c) 2025

accelerate==1.9.0
bitsandbytes==0.46.1
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.4
jira==3.8.0
joblib==1.5.1
numpy>=1.26.0
packaging==25.0
peft==0.16.0
pre_commit==4.2.0
prometheus_client==0.20.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
slack_bolt==1.23.0
slack_sdk==3.36.0
SQLAlchemy==2.0.41
starlette==0.47.2
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.3
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
